Move Assignments to SPI
