+++ presto-main/src/main/java/com/facebook/presto/sql/planner/optimizations/ApplyNodeUtil.java
@@ -13,7 +13,7 @@
  */
 package com.facebook.presto.sql.planner.optimizations;
 
-import com.facebook.presto.sql.planner.plan.Assignments;
+import com.facebook.presto.spi.plan.Assignments;
 import com.facebook.presto.sql.relational.OriginalExpressionUtils;
 import com.facebook.presto.sql.tree.ExistsPredicate;
 import com.facebook.presto.sql.tree.Expression;
