+++ presto-main/src/test/java/com/facebook/presto/sql/planner/iterative/rule/test/TestRuleTester.java
@@ -15,9 +15,9 @@ package com.facebook.presto.sql.planner.iterative.rule.test;
 
 import com.facebook.presto.matching.Captures;
 import com.facebook.presto.matching.Pattern;
+import com.facebook.presto.spi.plan.Assignments;
 import com.facebook.presto.spi.plan.PlanNode;
 import com.facebook.presto.sql.planner.iterative.Rule;
-import com.facebook.presto.sql.planner.plan.Assignments;
 import com.google.common.collect.ImmutableList;
 import org.testng.annotations.Test;
